Update realization display in case the action or the program name has been changed
