+++ services/src/main/java/org/exoplatform/addons/gamification/entities/domain/effective/GamificationActionsHistory.java
@@ -156,19 +156,19 @@ import org.exoplatform.commons.api.persistence.ExoEntity;
 )
 @NamedQuery(
     name = "GamificationActionsHistory.findAllAnnouncementByChallenge",
-    query = "SELECT DISTINCT a FROM GamificationActionsHistory a where a.ruleId = :challengeId order by a.id desc"
+    query = "SELECT a FROM GamificationActionsHistory a where a.ruleId = :challengeId order by a.id desc"
 )
 @NamedQuery(
     name = "GamificationActionsHistory.findAllAnnouncementByChallengeByEarnerType",
-    query = "SELECT DISTINCT a FROM GamificationActionsHistory a where a.ruleId = :challengeId AND a.earnerType = :earnerType order by a.id desc"
+    query = "SELECT a FROM GamificationActionsHistory a where a.ruleId = :challengeId AND a.earnerType = :earnerType order by a.id desc"
 )
 @NamedQuery(
     name = "GamificationActionsHistory.findAllAnnouncementByChallengeByDate",
-    query = "SELECT DISTINCT a FROM GamificationActionsHistory a where a.ruleId = :challengeId AND a.createdDate >= :fromDate AND a.createdDate < :toDate order by a.id desc"
+    query = "SELECT a FROM GamificationActionsHistory a where a.ruleId = :challengeId AND a.createdDate >= :fromDate AND a.createdDate < :toDate order by a.id desc"
 )
 @NamedQuery(
     name = "GamificationActionsHistory.findAllAnnouncementByChallengeByDateByEarnerType",
-    query = "SELECT DISTINCT a FROM GamificationActionsHistory a where a.ruleId = :challengeId AND a.createdDate >= :fromDate AND a.createdDate < :toDate AND a.earnerType = :earnerType order by a.id desc"
+    query = "SELECT a FROM GamificationActionsHistory a where a.ruleId = :challengeId AND a.createdDate >= :fromDate AND a.createdDate < :toDate AND a.earnerType = :earnerType order by a.id desc"
 )
 @NamedQuery(
     name = "GamificationActionsHistory.findMostRealizedRuleIds",
