+++ services/src/main/java/org/exoplatform/addons/gamification/rest/GamificationInformationsEndpoint.java
@@ -33,7 +33,7 @@ import org.apache.commons.lang3.StringUtils;
 
 import org.exoplatform.addons.gamification.IdentityType;
 import org.exoplatform.addons.gamification.entities.domain.effective.GamificationActionsHistory;
-import org.exoplatform.addons.gamification.service.dto.configuration.DomainDTO;
+import org.exoplatform.addons.gamification.rest.model.GamificationHistoryInfo;
 import org.exoplatform.addons.gamification.service.effective.GamificationService;
 import org.exoplatform.addons.gamification.service.mapper.DomainMapper;
 import org.exoplatform.addons.gamification.utils.Utils;
@@ -167,157 +167,5 @@ public class GamificationInformationsEndpoint implements ResourceContainer {
 
   private String getCurrentUsername() {
     return ConversationState.getCurrent().getIdentity().getUserId();
-  }
-
-  public static class GamificationHistoryInfo {
-    String         socialId;
-
-    boolean        isSpace;
-
-    String         avatarUrl;
-
-    String         remoteId;
-
-    String         fullname;
-
-    String         profileUrl;
-
-    String         createdDate;
-
-    long           globalScore;
-
-    String         actionTitle;
-
-    DomainDTO      domainDTO;
-
-    String         context;
-
-    long           actionScore;
-
-    private int    loadCapacity = 10;
-
-    private String receiver;
-
-    private String objectId;
-
-    public int getLoadCapacity() {
-      return loadCapacity;
-    }
-
-    public void setLoadCapacity(String loadCapacity) {
-      this.loadCapacity = Integer.parseInt(loadCapacity);
-    }
-
-    public String getAvatarUrl() {
-      return avatarUrl;
-    }
-
-    public void setAvatarUrl(String avatarUrl) {
-      this.avatarUrl = avatarUrl;
-    }
-
-    public String getRemoteId() {
-      return remoteId;
-    }
-
-    public void setRemoteId(String remoteId) {
-      this.remoteId = remoteId;
-    }
-
-    public String getFullname() {
-      return fullname;
-    }
-
-    public void setFullname(String fullname) {
-      this.fullname = fullname;
-    }
-
-    public String getProfileUrl() {
-      return profileUrl;
-    }
-
-    public void setProfileUrl(String profileUrl) {
-      this.profileUrl = profileUrl;
-    }
-
-    public String getSocialId() {
-      return socialId;
-    }
-
-    public void setSocialId(String socialId) {
-      this.socialId = socialId;
-    }
-
-    public boolean isSpace() {
-      return isSpace;
-    }
-
-    public void setSpace(boolean isSpace) {
-      this.isSpace = isSpace;
-    }
-
-    public long getGlobalScore() {
-      return globalScore;
-    }
-
-    public void setGlobalScore(long globalScore) {
-      this.globalScore = globalScore;
-    }
-
-    public String getActionTitle() {
-      return actionTitle;
-    }
-
-    public void setActionTitle(String actionTitle) {
-      this.actionTitle = actionTitle;
-    }
-
-    public DomainDTO getDomainDTO() {
-      return domainDTO;
-    }
-
-    public void setDomain(DomainDTO domainDTO) {
-      this.domainDTO = domainDTO;
-    }
-
-    public String getContext() {
-      return context;
-    }
-
-    public void setContext(String context) {
-      this.context = context;
-    }
-
-    public long getActionScore() {
-      return actionScore;
-    }
-
-    public void setActionScore(long actionScore) {
-      this.actionScore = actionScore;
-    }
-
-    public String getCreatedDate() {
-      return createdDate;
-    }
-
-    public void setCreatedDate(String createdDate) {
-      this.createdDate = createdDate;
-    }
-
-    public String getReceiver() {
-      return receiver;
-    }
-
-    public void setReceiver(String receiver) {
-      this.receiver = receiver;
-    }
-
-    public String getObjectId() {
-      return objectId;
-    }
-
-    public void setObjectId(String objectId) {
-      this.objectId = objectId;
-    }
   }
 }
