+++ services/src/main/resources/locale/addon/Gamification_en.properties
@@ -429,6 +429,7 @@ realization.label.lengthExceed=Exactly less than 1300 characters
 realization.label.save=Save
 realization.label.cancel=Cancel
 realization.label.sortBy=Sort By
+realization.label.previouslyNamed=Previously named
 
 gamification.type=Type
